Use proxy types package
