+++ packages/client/package.json
@@ -21,7 +21,7 @@
 	},
 	"packageManager": "yarn@4.1.0",
 	"devDependencies": {
-		"@open-frames/proxy": "workspace:*",
+		"@open-frames/proxy-types": "workspace:*",
 		"typescript": "^5.3.3",
 		"vitest": "^1.3.0"
 	}
