+++ packages/client/src/types.ts
@@ -1,4 +1,4 @@
-export type { GetMetadataResponse, OpenFrameButton, OpenFrameImage, OpenFrameResult, PostRedirectResponse } from '@open-frames/proxy';
+export type { GetMetadataResponse, OpenFrameButton, OpenFrameImage, OpenFrameResult, PostRedirectResponse } from '@open-frames/proxy-types';
 
 type JSONPrimitive = string | number | boolean | null;
 // eslint-disable-next-line no-use-before-define
