+++ packages/server/src/index.ts
@@ -2,7 +2,6 @@ import { CORS_HEADERS } from './constants.js';
 import { ErrorResponse } from './errors.js';
 import { handleGet, handleMedia, handlePost, handleRedirect } from './handlers.js';
 import { getRequestPath } from './utils.js';
-export * from './types.js';
 
 export async function handleRequest(req: Request): Promise<Response> {
 	try {
