+++ packages/types/package.json
@@ -1,0 +1,24 @@
+{
+	"name": "@open-frames/proxy-types",
+	"version": "0.0.1",
+	"types": "dist/index.d.ts",
+	"packageManager": "yarn@4.1.0",
+	"type": "module",
+	"publishConfig": {
+		"access": "public",
+		"provenance": true
+	},
+	"scripts": {
+		"build": "tsc"
+	},
+	"engines": {
+		"node": ">=18"
+	},
+	"devDependencies": {
+		"@types/node": "^20.11.19",
+		"typescript": "^5.3.3"
+	},
+	"peerDependencies": {
+		"typescript": "^5.3.3"
+	}
+}
